+++ src/poll.cpp
@@ -124,7 +124,7 @@ void zmq::poll_t::loop ()
     while (!stopping) {
 
         //  Execute any due timers.
-        uint64_t timeout = execute_timers ();
+        int timeout = (int) execute_timers ();
 
         //  Wait for events.
         int rc = poll (&pollset [0], pollset.size (), timeout ? timeout : -1);
